made small refactor of styles and up version
