+++ src/components/familyInfoPage/DependantsInfoItem.js
@@ -1,7 +1,7 @@
 import { Text, View, TouchableOpacity, StyleSheet } from "react-native";
 import PersonIconSvg from "../../assets/icons/familyInfoIcons/person.svg";
 import ArrowRightSvg from "../../assets/icons/familyInfoIcons/vectorRight.svg";
-import mainStyles from "./stylesFamilyInfoItem";
+import mainStyles from "./FamilyInfoItemStyles";
 import TrashBasketSvg from "../../assets/icons/familyInfoIcons/trashBasket.svg";
 import { THEME } from "../../styles/theme";
 import CrossSvg from "../../assets/icons/familyInfoIcons/cross.svg";
